+++ frontend-react/src/utils/SessionStorageTools.ts
@@ -16,6 +16,9 @@ export enum GLOBAL_STORAGE_KEYS {
     ORGANIZATION_OVERRIDE = "global-organization-override",
 }
 
+/* feature flags are just and array of strings saved into a single localStorage variable */
+const FEATURE_FLAG_LOCALSTORAGE_KEY = "featureFlags";
+
 const fetchJsonFromSession = (storageKey: string) => {
     const storedString = sessionStorage.getItem(storageKey);
     if (!storedString) {
@@ -78,4 +81,13 @@ export function storeSessionMembershipState(membershipState: string) {
         GLOBAL_STORAGE_KEYS.MEMBERSHIP_STATE,
         membershipState
     );
+}
+
+export function getSavedFeatureFlags(): string[] {
+    const saved =
+        window.localStorage.getItem(FEATURE_FLAG_LOCALSTORAGE_KEY) || "";
+    if (saved === "") {
+        return [];
+    }
+    return saved.split("\t");
 }
